fix(core): invert stub/real split to fix Babel build error
Previous approach tried to redirect webpack away from inter.ts (which
imports next/font/google) using NormalModuleReplacementPlugin. This
did not work reliably because webpack follows require() statically
before the plugin can intercept.

New approach inverts the responsibility:

- src/fonts/inter.ts  →  Babel-safe stub (exports null by default)
- src/fonts/inter.optimized.ts  →  real next/font/google module

Since inter.ts is now the default module webpack compiles, stores that
use Babel never encounter the "next/font requires SWC" error.
inter.optimized.ts is only pulled into the module graph when
experimental.optimizedFonts === true via NormalModuleReplacementPlugin,
at which point SWC is required and expected.
