--- packages/core/src/fonts/inter.stub.ts
@@ -1,4 +1,0 @@
-// Stub used when experimental.optimizedFonts is false.
-// Webpack aliases src/fonts/inter to this file so that inter.ts
-// (which imports next/font/google and requires SWC) is never compiled.
-export const inter = null
