+++ packages/core/test/fonts/inter.test.ts
@@ -13,16 +13,23 @@ vi.mock('next/font/google', () => ({
   },
 }))
 
-describe('inter font module', () => {
-  it('exports an inter font object with the expected variable', async () => {
+describe('inter.ts (default stub — Babel-safe)', () => {
+  it('exports null so builds with Babel never load next/font/google', async () => {
     const { inter } = await import('src/fonts/inter')
+    expect(inter).toBeNull()
+  })
+})
+
+describe('inter.optimized.ts (SWC-only, loaded when optimizedFonts: true)', () => {
+  it('exports an inter font object with the expected variable', async () => {
+    const { inter } = await import('src/fonts/inter.optimized')
 
     expect(inter).toBeDefined()
     expect(inter.variable).toBe('--font-inter')
   })
 
   it('configures only latin and latin-ext subsets', async () => {
-    await import('src/fonts/inter')
+    await import('src/fonts/inter.optimized')
 
     expect(mockInter).toHaveBeenCalledWith(
       expect.objectContaining({
@@ -32,7 +39,7 @@ describe('inter font module', () => {
   })
 
   it('configures display: swap for performance', async () => {
-    await import('src/fonts/inter')
+    await import('src/fonts/inter.optimized')
 
     expect(mockInter).toHaveBeenCalledWith(
       expect.objectContaining({
